requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
black==21.6b0
bleach==3.3.0
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
Cython
decorator==5.0.9
defusedxml==0.7.1
entrypoints==0.3
fastapi==0.65.2
fastapi-cache2==0.1.3.4
h11==0.12.0
idna==2.10
ipykernel==5.5.5
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
murmurhash==1.0.5
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.22.2
opencv-python==4.5.5.62
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.8.1
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
preshed==3.0.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
pydantic==1.7.4
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2021.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.8.3
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
six==1.16.0
sklearn==0.0
smart-open==5.1.0
spacy==3.0.6
spacy-legacy==3.0.6
srsly==2.4.1
SQLAlchemy==1.4.31
starlette==0.14.2
terminado==0.10.1
testpath==0.5.0
thinc==8.0.6
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.1
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
